refactor: abstract "getDeepElementFromPoint" into shared package
