+++ packages/overlay/src/overlay-stack.ts
@@ -22,6 +22,7 @@ import {
     parentOverlayOf,
 } from './overlay-utils.js';
 import { OverlayCloseEvent } from './overlay-events.js';
+import { getDeepElementFromPoint } from '@spectrum-web-components/shared/src/get-deep-element-from-point.js';
 
 function isLeftClick(event: MouseEvent): boolean {
     return event.button === 0;
@@ -184,19 +185,9 @@ export class OverlayStack {
         event.stopPropagation();
         event.preventDefault();
         await this.closeTopOverlay();
-        let target = document.elementFromPoint(event.clientX, event.clientY);
-        while (target?.shadowRoot) {
-            const innerTarget = (
-                target.shadowRoot as unknown as {
-                    elementFromPoint: (x: number, y: number) => Element | null;
-                }
-            ).elementFromPoint(event.clientX, event.clientY);
-            if (!innerTarget || innerTarget === target) {
-                break;
-            }
-            target = innerTarget;
-        }
-        target?.dispatchEvent(new MouseEvent('contextmenu', event));
+        getDeepElementFromPoint(event.clientX, event.clientY)?.dispatchEvent(
+            new MouseEvent('contextmenu', event)
+        );
     };
 
     private get document(): Document {
